% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CFP_data}
\alias{CFP_data}
\title{Corporate financial performance (CFP) data for selected African companies}
\format{
A data frame with 1195 rows, 12 variables:

\describe{
\item{Company}{Company name}
\item{Country}{Country of origin}
\item{Multinational}{Whether multinational}
\item{Sector}{Economic sector }
\item{Listing Age}{Years since listing}
\item{Incorp. Age}{Years since incorporation}
\item{Size}{Firm size in USD}
\item{Return on Assets (lagged)}{Return on assest}
\item{Return on Equity (lagged)}{Return on equity}
\item{Social Comm.}{Existence of social and ethics commitee}
\item{Year}{Reporting year}
\item{Net Profit}{Reported annual net profit}
}
}
\source{
Data gathered from published company annual reports
}
\usage{
CFP_data
}
\description{
A dataset containing variables on CFP and company characteristics
for 238 African companies spanning a period of 5 years (2014 to 2018)
}
\keyword{datasets}
