% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_indices}
\alias{country_indices}
\title{Country characteristics data}
\format{
30 rows, 7 variables

\describe{
\item{Country}{Country of origin}
\item{Population}{Pop}
\item{GDP $}{GDP in USD}
\item{GPD/Capita}{Per capita GDP in USD}
\item{Life expentancy}{Av. life expectancy at birth}
\item{Corruption index}{Corruption Index}
\item{Year}{Reporting year}
}
}
\source{
Data gathered from World Bank
}
\usage{
country_indices
}
\description{
A dataset containing selected socioeconomic and demographic
characteristics for 6 African countries for the
period 2014 to 2018)
}
\keyword{datasets}
