% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CSR_data}
\alias{CSR_data}
\title{Corporate social performance (CSP) data for selected African companies}
\format{
A data frame with 1195 rows, 27 variables:

\describe{
\item{Company}{Compnay name}
\item{Year}{Reporting year}
\item{Country}{Listing Country}
\item{Sector}{Industrial Sector}
\item{Education CSR}{Education interventions}
\item{Health CSR}{Health  interventions}
\item{Culture CSR}{Cultural activitess}
\item{Economic CSR}{Economic activities}
\item{Agric. CSR}{Agric support}
\item{Soc. Mobilization CSR}{Social mobilization activities}
\item{Emp. Safety CSR}{Improving worker safety}
\item{Emp. Diversity CSR}{Workforce diversity}
\item{Emp. Training CSR}{Workforce capacity building}
\item{Emp. Benefits CSR}{Employee benefits}
\item{Emp. Monitoring CSR}{Workplace statistics tracking}
\item{Emp. Satisfaction CSR}{Worker satisfaction}
\item{Prod. Info CSR}{Product information dissemination}
\item{Prod. Quality CSR}{Product quality}
\item{Energy Conserv. CSR}{Energy conservation activities}
\item{Energy Policy CSR}{Energy policy}
\item{Emissions Policy CSR}{Emissions policy}
\item{Emissions Mitigation CSR}{Activities to mitigate emissions}
\item{Pollution CSR}{Pollution reduction}
\item{Conservation CSR}{Env. conservation activities}
\item{Env. Policy CSR}{Env. policy}
\item{Water Policy CSR}{Water conservation policy}
\item{Other CSR}{Other CSR activities}
}
}
\source{
Data gathered from published company annual reports
}
\usage{
CSR_data
}
\description{
Dataset containing variables on 23 CSP domains, and 4 company characteristics
for 238 African companies spanning a period of 5 years (2014 to 2018)
}
\keyword{datasets}
