% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exchange_rates}
\alias{exchange_rates}
\title{Exchange rate data characteristics data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1190 rows and 3 columns.
}
\source{
Data gathered from online
}
\usage{
exchange_rates
}
\description{
A dataset containing exchange rate data for 6 African currencies
from 2014 to 2018
}
\details{
\describe{
\item{Company}{Company name}
\item{ex_rate}{Exchange rate rel to USD}
\item{Year}{Reporting year}
}
}
\keyword{datasets}
