% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pca_ind_scores}
\alias{pca_ind_scores}
\title{Corporate social performance (CSP) scores}
\format{
A data frame with 1195 rows, 1 variable:
}
\source{
Data gathered from published company annual reports
}
\usage{
pca_ind_scores
}
\description{
A dataset containing CSP scores obtained from principal component analysis
of 23 CSP variables for 238 African companies spanning a period of 5 years
(2014 to 2018)
}
\keyword{datasets}
